refactor: 백엔드 코드 모듈화 및 http method 별 분리
