+++ src/backend/auth/index.ts
@@ -1,0 +1,7 @@
+export * from '@/backend/auth/httpMethods/GET';
+export * from '@/backend/auth/httpMethods/POST';
+export * from '@/backend/auth/httpMethods/DELETE';
+
+export type * from '@/backend/auth/httpMethods/GET';
+export type * from '@/backend/auth/httpMethods/POST';
+export type * from '@/backend/auth/httpMethods/DELETE';
