+++ src/backend/utils/cookieUtils.ts
